requirements.txt
openai>=2.0.0
python-dotenv>=1.0.0
Flask>=2.1.0
Flask-Cors>=3.0.10
